feat: styling changes on header bar
