+++ src/Shared/ContextMenu.vue
@@ -3,12 +3,12 @@
     <div>
       <MenuButton
         :disabled="!hasOptions"
-        class="inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
+        class="inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm text-gray-700 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
       >
         {{ anchor }}
         <Icon
           v-if="hasOptions"
-          class="h-5 w-5 text-gray-400"
+          class="h-5 w-5 text-gray-700"
           aria-hidden="true"
           name="chevron-down"
         />
