--- src/Shared/NavigationBar.story.vue
@@ -1,162 +1,0 @@
-<template>
-  <Story title="Navigation Bar" group="shared">
-    <Variant title="SingleStory SingleLanguage Editor" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="MultiStory SingleLanguage Editor" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="MultiStory MultiLanguage Editor" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="MultiStory MultiLanguage Admin" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="SingleStory MultiLanguage Editor" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="SingleStory MultiLanguage Admin" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="SingleStory SingleLanguage Admin" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-    <Variant title="MultiStory SingleLanguage Admin" :setup-app="loadData">
-      <NavigationBar />
-    </Variant>
-  </Story>
-</template>
-
-<script setup lang="ts">
-import NavigationBar from './NavigationBar.vue';
-import { LanguageSpecification } from 'src/Shared/interfaces';
-import { useSharedStore } from '../store';
-import type { Vue3StorySetupHandler } from '@histoire/plugin-vue';
-import { sharedProps, meta } from '../helpers/mocks';
-
-import { useForm } from '@inertiajs/vue3';
-
-const loadData: Vue3StorySetupHandler = ({ variant }) => {
-  const shared = useSharedStore();
-
-  shared.setFromProps(sharedProps);
-  // shared.meta.logo =
-  //   'https://res.cloudinary.com/theword121/image/upload/v1687245360/episodes/viseg2hegowcrapio6pt.svg';
-  // shared.meta.name = 'CMS UI';
-
-  if (variant?.title == 'SingleStory SingleLanguage Editor') {
-    shared.meta = meta;
-    shared.languages = [
-      {
-        locale: 'en',
-        language: 'English',
-        languageDirection: 'ltr',
-      },
-    ];
-    shared.stories = ['John'];
-    shared.user.isAdmin = false;
-    shared.user.isManager = false;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  if (variant?.title == 'MultiStory SingleLanguage Editor') {
-    shared.meta = meta;
-    shared.languages = [
-      {
-        locale: 'en',
-        language: 'English',
-        languageDirection: 'ltr',
-      },
-    ];
-    shared.user.isAdmin = false;
-    shared.user.isManager = false;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  if (variant?.title == 'MultiStory MultiLanguage Editor') {
-    shared.meta = meta;
-    shared.user.isAdmin = false;
-    shared.user.isManager = false;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-
-  if (variant?.title == 'MultiStory MultiLanguage Admin') {
-    shared.meta = meta;
-    shared.user.isAdmin = true;
-    shared.user.isManager = true;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  if (variant?.title == 'SingleStory MultiLanguage Editor') {
-    shared.meta = meta;
-    shared.stories = ['John'];
-    shared.user.isAdmin = false;
-    shared.user.isManager = false;
-    shared.user.role = 'editor';
-    shared.user.language = '*';
-  }
-
-  if (variant?.title == 'SingleStory MultiLanguage Admin') {
-    shared.meta = meta;
-    shared.stories = ['John'];
-    shared.user.isAdmin = true;
-    shared.user.isManager = true;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  if (variant?.title == 'SingleStory SingleLanguage Admin') {
-    shared.meta = meta;
-    shared.languages = [
-      {
-        locale: 'en',
-        language: 'English',
-        languageDirection: 'ltr',
-      },
-    ];
-    shared.stories = ['John'];
-    shared.user.isAdmin = true;
-    shared.user.isManager = false;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  // %%%%%%%%%%%%%%%%
-
-  if (variant?.title == 'MultiStory SingleLanguage Admin') {
-    shared.meta = meta;
-    shared.languages = [
-      {
-        locale: 'en',
-        language: 'English',
-        languageDirection: 'ltr',
-      },
-    ];
-    shared.user.isAdmin = true;
-    shared.user.isManager = true;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-  if (variant?.title == 'Admin All') {
-    shared.meta = meta;
-    shared.user.isAdmin = true;
-    shared.user.isManager = true;
-    shared.user.role = 'admin';
-    shared.user.language = '*';
-  }
-
-  interface Form {
-    language: string | null;
-    story: string | null;
-  }
-
-  const form = useForm({
-    language: shared.language.language,
-    story: null,
-  } as Form);
-};
-const language: LanguageSpecification = {
-  locale: 'es',
-  language: 'Espanol',
-  languageDirection: 'ltr',
-};
-</script>
