+++ src/Shared/NavigationBar.vue
@@ -10,23 +10,24 @@
         }}</Link>
         <!-- eslint-disable vue/valid-v-model -->
         <DropDown
-          v-if="isMultiLingual"
           v-model="(form.language as string)"
           :options="(shared.languages.map((l) => l.language) as string[])"
           :is-read-only="!shared.user.isManager"
           @change="onLanguage"
         ></DropDown>
+      </div>
+      <div class="flex items-center space-x-6 md:ml-24">
         <ContextMenu
-          v-if="isMultiStory"
           :options="(shared.stories as string[])"
           :anchor="shared.meta.storyType"
           @select="onStory"
         ></ContextMenu>
-
-        <Link v-if="!isMultiStory" class="px-2 py-3 hover:text-gray-700" href="/"
-          >{{ shared.meta.storyType }}s</Link
+        <Link
+          v-if="shared.user.isAdmin"
+          class="px-2 py-3 hover:text-gray-700"
+          href="/page"
+          >Pages</Link
         >
-        <Link class="px-2 py-3 hover:text-gray-700" href="/page">Pages</Link>
         <Link
           v-if="shared.user.isAdmin"
           class="px-2 py-3 hover:text-gray-700"
@@ -90,7 +91,6 @@
 
 <script setup lang="ts">
 import { Link, useForm } from '@inertiajs/vue3';
-import { computed } from 'vue';
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/vue';
 import DropDown from './DropDown.vue';
 import ContextMenu from './ContextMenu.vue';
@@ -122,6 +122,4 @@ const onStory = async (story: string) => {
 };
 
 const signOut = () => (window.location.href = '/logout');
-const isMultiLingual = computed(() => shared.languages.length > 1);
-const isMultiStory = computed(() => shared.stories.length > 1);
 </script>
